# src/token/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "XRPL Token Insurance Demo API",
    "version": "0.4.0",
    "description": "REST API for XRPL Token Insurance Demo platform - stablecoin (JPYN) issuance with fiat↔token exchange"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Google OAuth2 ID Token"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string" }
        }
      },
      "BankAccount": {
        "type": "object",
        "required": ["bankCode", "branchCode", "accountNumber", "accountHolder", "label"],
        "properties": {
          "bankCode": { "type": "string" },
          "branchCode": { "type": "string" },
          "accountNumber": { "type": "string" },
          "accountHolder": { "type": "string" },
          "label": { "type": "string" },
          "createdAt": { "type": "string" }
        }
      },
...
      "WhitelistAddress": {
        "type": "object",
        "properties": {
          "address": { "type": "string" },
          "label": { "type": "string" },
          "createdAt": { "type": "string" }
        }
      },
...
      "FiatWithdrawalResult": {
        "type": "object",
        "properties": {
          "amount": { "type": "number" },
          "destination": {
            "type": "object",
            "properties": {
              "bankCode": { "type": "string" },
              "branchCode": { "type": "string" },
              "accountNumber": { "type": "string" },
              "accountHolder": { "type": "string" }
            }
          },
          "txReference": { "type": "string" }
        }
      },
      "XrpWithdrawalResult": {
        "type": "object",
        "properties": {
          "tokenId": { "type": "string" },
          "amount": { "type": "number" },
          "destinationAddress": { "type": "string" },
          "xrplTxHash": { "type": "string" }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/whitelist/xrp": {
      "get": {
        "operationId": "getXrpWhitelist",
        "summary": "Get XRP withdrawal address whitelist",
        "tags": ["Whitelist"],
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "XRP whitelist",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/WhitelistAddress" }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "addXrpWhitelist",
        "summary": "Add XRP address to whitelist",
        "tags": ["Whitelist"],
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["address", "label"],
                "properties": {
                  "address": { "type": "string" },
                  "label": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Address added to whitelist",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/WhitelistAddress" }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/api/v1/whitelist/xrp/{address}": {
      "delete": {
        "operationId": "removeXrpWhitelist",
        "summary": "Remove XRP address from whitelist",
        "tags": ["Whitelist"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Address removed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "ok" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Address not found in whitelist",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/api/v1/whitelist/bank": {
      "get": {
        "operationId": "getBankWhitelist",
        "summary": "Get bank withdrawal whitelist",
        "tags": ["Whitelist"],
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Bank whitelist",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/BankAccount" }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "addBankWhitelist",
        "summary": "Add bank account to whitelist",
        "tags": ["Whitelist"],
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["bankCode", "branchCode", "accountNumber", "accountHolder", "label"],
                "properties": {
                  "bankCode": { "type": "string" },
                  "branchCode": { "type": "string" },
                  "accountNumber": { "type": "string" },
                  "accountHolder": { "type": "string" },
                  "label": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Bank account added to whitelist",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BankAccount" }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/api/v1/whitelist/bank/{id}": {
      "delete": {
        "operationId": "removeBankWhitelist",
        "summary": "Remove bank account from whitelist (id = branchCode-accountNumber)",
        "tags": ["Whitelist"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "Format: branchCode-accountNumber"
          }
        ],
        "responses": {
          "200": {
            "description": "Bank account removed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "ok" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Bank account not found in whitelist",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/api/v1/withdraw/fiat": {
      "post": {
        "operationId": "withdrawFiat",
        "summary": "Withdraw fiat to bank account",
        "tags": ["Withdrawal"],
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["amount", "bankAccount"],
                "properties": {
                  "amount": { "type": "number", "minimum": 1 },
                  "bankAccount": { "$ref": "#/components/schemas/BankAccount" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Fiat withdrawal result",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/FiatWithdrawalResult" }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/api/v1/withdraw/xrp": {
      "post": {
        "operationId": "withdrawXrp",
        "summary": "Withdraw XRP token to external address",
        "tags": ["Withdrawal"],
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["tokenId", "tokenAmount", "destinationAddress"],
                "properties": {
                  "tokenId": { "type": "string" },
                  "tokenAmount": { "type": "number", "minimum": 1 },
                  "destinationAddress": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "XRP withdrawal result",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/XrpWithdrawalResult" }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
  }
}